# public/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Luna AI API",
    "version": "1.0.0",
    "description": "API Documentation for Luna AI Streamer - AI-powered streamer with VTube Studio integration, emotion system, and Rock Paper Scissors game",
    "contact": {
      "name": "Luna AI Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8787",
      "description": "Local development server"
    },
    {
      "url": "https://your-domain.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/luna/expression": {
      "get": {
        "tags": ["VTS"],
        "summary": "Test emotion expression",
        "description": "Trigger an emotion expression in VTube Studio",
        "parameters": [
          {
            "name": "emo",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["angry", "sad", "sleepy", "hype", "soft", "clear"]
            },
            "description": "Emotion to trigger"
          }
        ],
        "responses": {
          "200": {
            "description": "Expression triggered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "emotion": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "AdminSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-secret",
        "description": "Admin secret key (set ADMIN_SECRET in .env)"
      },
      "PurchaseSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-purchase-secret",
        "description": "Purchase webhook secret (set PURCHASE_SECRET in .env)"
      }
    }
  }
}